Change the type of value for the queries, it wasn't working

+++ src/main/java/JEstebanC/FastFoodApp/dto/ReportClientDTO.java
@@ -5,6 +5,8 @@ package JEstebanC.FastFoodApp.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author Juan Esteban Castaño Holguin castanoesteban9@gmail.com 2022-03-19
  */
@@ -14,5 +16,5 @@ public class ReportClientDTO {
 	private Long idUser;
 	private String username;
 	private String urlImage;
-	private int total;
+	private BigInteger total;
 }
