+++ src/main/java/JEstebanC/FastFoodApp/dto/ReportProductDTO.java
@@ -4,6 +4,8 @@
 package JEstebanC.FastFoodApp.dto;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author Juan Esteban Castaño Holguin castanoesteban9@gmail.com 2022-03-16
  */
@@ -13,5 +15,5 @@ public class ReportProductDTO {
 	private Long idProduct;
 	private String name;
 	private int amount;
-	private int total;
+	private BigInteger total;
 }
