+++ src/main/java/JEstebanC/FastFoodApp/dto/ReportSalesDTO.java
@@ -5,6 +5,8 @@ package JEstebanC.FastFoodApp.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author Juan Esteban Castaño Holguin castanoesteban9@gmail.com 2022-03-21
  */
@@ -13,5 +15,5 @@ public class ReportSalesDTO {
 
 	private Long idBill;
 
-	private int total;
+	private BigInteger total;
 }
