+++ src/main/java/JEstebanC/FastFoodApp/dto/ReportSalesMonthlyDTO.java
@@ -2,11 +2,13 @@ package JEstebanC.FastFoodApp.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author Juan Esteban Castaño Holguin castanoesteban9@gmail.com 6/13/2022
  */
 @Data
 public class ReportSalesMonthlyDTO {
-    private Double month;
-    private int total;
+    private Long month;
+    private BigInteger total;
 }
